+++ packages/components/src/Avatar/AvatarUser.tsx
@@ -89,6 +89,8 @@ export const AvatarUser = styled(AvatarLayout).attrs((props) => ({
   size: props.size || 'small',
 }))`
   ${avatarCSS}
+  ${({ role }) =>
+    role === 'button' && 'cursor: pointer;'}
 
   background: currentColor;
   position: relative;