+++ packages/components/src/Button/ButtonItem.tsx
@@ -91,6 +91,7 @@ export const ButtonItem = styled(ButtonLayout)`
   color: ${({ theme }) => theme.colors.text3};
   cursor: pointer;
   display: inline-flex;
+  font-family: ${({ theme }) => theme.fonts.brand};
   font-size: ${({ theme }) => theme.fontSizes.small};
   height: ${inputHeight};
   justify-content: center;